+++ src/components/LayerListItem.tsx
@@ -5,16 +5,19 @@ import {MdContentCopy, MdVisibility, MdVisibilityOff, MdDelete} from 'react-icon
 import { IconContext } from 'react-icons'
 
 import IconLayer from './IconLayer'
-import {SortableElement, SortableHandle} from 'react-sortable-hoc'
+import {useSortable, type UseSortableArguments} from '@dnd-kit/sortable'
+import {CSS} from '@dnd-kit/utilities'
 
 
 type DraggableLabelProps = {
   layerId: string
   layerType: string
+  listeners: any
+  attributes: Record<string, any>
 };
 
-const DraggableLabel = SortableHandle((props: DraggableLabelProps) => {
-  return <div className="maputnik-layer-list-item-handle">
+function DraggableLabel(props: DraggableLabelProps) {
+  return <div className="maputnik-layer-list-item-handle" {...props.attributes} {...(props.listeners as any)}>
     <IconLayer
       className="layer-handle__icon"
       type={props.layerType}
@@ -23,7 +26,7 @@ const DraggableLabel = SortableHandle((props: DraggableLabelProps) => {
       {props.layerId}
     </button>
   </div>
-});
+}
 
 type IconActionProps = {
   action: string
@@ -82,55 +85,69 @@ type LayerListItemProps = {
   onLayerVisibilityToggle?(...args: unknown[]): unknown
 };
 
-class LayerListItem extends React.Component<LayerListItemProps> {
-  static defaultProps = {
-    isSelected: false,
-    visibility: 'visible',
-    onLayerCopy: () => {},
-    onLayerDestroy: () => {},
-    onLayerVisibilityToggle: () => {},
-  }
+function LayerListItem(props: LayerListItemProps) {
+  const {
+    attributes,
+    listeners,
+    setNodeRef,
+    transform,
+    transition,
+  } = useSortable({id: props.layerId} as UseSortableArguments);
 
-  render() {
-    const visibilityAction = this.props.visibility === 'visible' ? 'show' : 'hide';
+  const style = {
+    transform: CSS.Transform.toString(transform),
+    transition,
+  } as React.CSSProperties;
 
-    return <IconContext.Provider value={{size: '14px'}}>
+  const visibilityAction = props.visibility === 'visible' ? 'show' : 'hide';
+
+  return (
+    <IconContext.Provider value={{size: '14px'}}>
       <li
-        id={this.props.id}
-        key={this.props.layerId}
-        onClick={_e => this.props.onLayerSelect(this.props.layerIndex)}
-        data-wd-key={"layer-list-item:"+this.props.layerId}
+        ref={setNodeRef}
+        style={style}
+        id={props.id}
+        key={props.layerId}
+        onClick={_e => props.onLayerSelect(props.layerIndex)}
+        data-wd-key={"layer-list-item:"+props.layerId}
         className={classnames({
           "maputnik-layer-list-item": true,
-          "maputnik-layer-list-item-selected": this.props.isSelected,
-          [this.props.className!]: true,
-        })}>
-        <DraggableLabel {...this.props} />
+          "maputnik-layer-list-item-selected": props.isSelected,
+          [props.className!]: true,
+        })}
+      >
+        <DraggableLabel layerId={props.layerId} layerType={props.layerType} listeners={listeners} attributes={attributes} />
         <span style={{flexGrow: 1}} />
         <IconAction
-          wdKey={"layer-list-item:"+this.props.layerId+":delete"}
+          wdKey={"layer-list-item:"+props.layerId+":delete"}
           action={'delete'}
           classBlockName="delete"
-          onClick={_e => this.props.onLayerDestroy!(this.props.layerIndex)}
+          onClick={_e => props.onLayerDestroy!(props.layerIndex)}
         />
         <IconAction
-          wdKey={"layer-list-item:"+this.props.layerId+":copy"}
+          wdKey={"layer-list-item:"+props.layerId+":copy"}
           action={'duplicate'}
           classBlockName="duplicate"
-          onClick={_e => this.props.onLayerCopy!(this.props.layerIndex)}
+          onClick={_e => props.onLayerCopy!(props.layerIndex)}
         />
         <IconAction
-          wdKey={"layer-list-item:"+this.props.layerId+":toggle-visibility"}
+          wdKey={"layer-list-item:"+props.layerId+":toggle-visibility"}
           action={visibilityAction}
           classBlockName="visibility"
           classBlockModifier={visibilityAction}
-          onClick={_e => this.props.onLayerVisibilityToggle!(this.props.layerIndex)}
+          onClick={_e => props.onLayerVisibilityToggle!(props.layerIndex)}
         />
       </li>
     </IconContext.Provider>
-  }
+  );
 }
 
-const LayerListItemSortable = SortableElement<LayerListItemProps>((props: LayerListItemProps) => <LayerListItem {...props} />);
+LayerListItem.defaultProps = {
+  isSelected: false,
+  visibility: 'visible',
+  onLayerCopy: () => {},
+  onLayerDestroy: () => {},
+  onLayerVisibilityToggle: () => {},
+};
 
-export default LayerListItemSortable;
+export default LayerListItem;